+++ src/qemu/qemu_driver.c
@@ -5440,7 +5440,7 @@ static int
 qemudCanonicalizeMachineDirect(virDomainDefPtr def, char **canonical)
 {
     virCapsGuestMachinePtr *machines = NULL;
-    int i, nmachines = 0;
+    size_t i, nmachines = 0;
 
     /* XXX we should be checking emulator capabilities and pass them instead
      * of NULL so that -nodefconfig or -no-user-config is properly added when
